1.调整底层宽带指令发送接口 2.调整底层OTT指令发送接口 3.实现OTT_MOBILE接口 4.修改OTT和BAND的指令发送JOB 5.修改authCom*pent中指令时间的时分秒问题

+++ ycsoft-lib/src/main/java/com/ycsoft/beans/core/job/JBandCommand.java
@@ -27,7 +27,7 @@ public class JBandCommand implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 8914092755631544635L;
-	private Integer transnum ;	
+	private Long transnum ;	
 	private Integer job_id ;	
 	private String supplyier_id ;	
 	private String server_id ;	
@@ -56,11 +56,11 @@ public class JBandCommand implements Serializable {
 	
 	
 	// transnum getter and setter
-	public Integer getTransnum(){
+	public Long getTransnum(){
 		return this.transnum ;
 	}
 	
-	public void setTransnum(Integer transnum){
+	public void setTransnum(Long transnum){
 		this.transnum = transnum ;
 	}
